gcs_to_bq/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile gcs_to_bq/requirements.in
#
cachetools==4.1.1         # via google-auth
certifi==2020.6.20        # via requests
cffi==1.14.2              # via google-crc32c
chardet==3.0.4            # via requests
google-api-core==1.22.2   # via google-cloud-bigquery, google-cloud-core
google-auth==1.21.2       # via google-api-core, google-cloud-storage
google-cloud-bigquery==1.27.2  # via -r gcs_to_bq/requirements.in
google-cloud-core==1.4.1  # via google-cloud-bigquery, google-cloud-storage
google-cloud-storage==1.31.0  # via -r gcs_to_bq/requirements.in
google-cloud==0.34.0      # via -r gcs_to_bq/requirements.in
google-crc32c==1.0.0      # via google-resumable-media
google-resumable-media==1.0.0  # via google-cloud-bigquery, google-cloud-storage
googleapis-common-protos==1.52.0  # via google-api-core
idna==2.10                # via requests
numpy==1.19.2             # via pandas
pandas==1.1.2             # via -r gcs_to_bq/requirements.in
protobuf==3.13.0          # via google-api-core, googleapis-common-protos
pyasn1-modules==0.2.8     # via google-auth
pyasn1==0.4.8             # via pyasn1-modules, rsa
pycparser==2.20           # via cffi
python-dateutil==2.8.1    # via pandas
pytz==2020.1              # via google-api-core, pandas
requests==2.24.0          # via google-api-core
rsa==4.6                  # via google-auth
six==1.15.0               # via google-api-core, google-auth, google-cloud-bigquery, google-resumable-media, protobuf, python-dateutil
urllib3==1.25.10          # via requests
xlrd==1.2.0               # via -r gcs_to_bq/requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
